backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
pydantic==2.10.3
python-dotenv==1.0.1
pydantic-settings==2.7.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
requests==2.31.0
cloudinary==1.41.0
